+++ Reflection/src/utopia/reflection/component/reach/hierarchy/ComponentHierarchy.scala
@@ -1,28 +1,50 @@
 package utopia.reflection.component.reach.hierarchy
 
-import utopia.flow.datastructure.mutable.PointerWithEvents
-import utopia.flow.event.{ChangeEvent, ChangeListener}
-import utopia.genesis.shape.shape2D.{Bounds, Point, Vector2D}
+import utopia.flow.event.Changing
+import utopia.genesis.shape.shape2D.{Bounds, Vector2D}
+import utopia.reflection.component.reach.template.ReachComponentLike
+import utopia.reflection.container.swing.ReachCanvas
 import utopia.reflection.text.Font
+
+import scala.annotation.tailrec
 
 /**
   * Represents a sequence of components that forms a linear hierarchy
   * @author Mikko Hilpinen
   * @since 3.10.2020, v2
   */
-class ComponentHierarchy(top: HierarchyTop, blocks: Vector[HierarchyBlock])
+trait ComponentHierarchy
 {
-	// ATTRIBUTES	----------------------
+	// ABSTRACT	--------------------------
 	
-	private lazy val statusListener = new LinkChangeListener()
+	/**
+	  * @return The next "block" in this hierarchy (either Left: Canvas at the top or
+	  *         Right: an intermediate block + a component associated with that block)
+	  */
+	def parent: Either[ReachCanvas, (ComponentHierarchy, ReachComponentLike)]
+	
+	/**
+	  * @return A pointer that shows whether this hierarchy is currently active / linked to the top window. Should
+	  *         take into account possible parent state.
+	  */
+	def linkPointer: Changing[Boolean]
+	
+	/**
+	  * @return Whether the link between this component and the parent component should be considered active
+	  */
+	def isThisLevelLinked: Boolean
 	
 	
 	// COMPUTED	--------------------------
 	
 	/**
-	  * @return A pointer that contains whether this hierarchy reaches the top or not
+	  * @return The canvas at the top of this hierarchy
 	  */
-	def linkPointer = statusListener.pointer.view
+	def top: ReachCanvas = parent match
+	{
+		case Left(canvas) => canvas
+		case Right((block, _)) => block.top
+	}
 	
 	/**
 	  * @return Whether this hierarchy currently reaches the top component without any broken links
@@ -32,39 +54,66 @@ class ComponentHierarchy(top: HierarchyTop, blocks: Vector[HierarchyBlock])
 	/**
 	  * @return The window that contains this component hierarchy. None if not connected to a window.
 	  */
-	def parentWindow = top.canvas.parentWindow
+	def parentWindow = top.parentWindow
 	
 	/**
 	  * @return A modifier used when calculating the position of the bottom component (outside this hierarchy)
 	  *         relative to hierarchy top
 	  */
-	def positionToTopModifier = blocks.foldLeft(Vector2D.zero) { (pos, block) =>
-		pos + block.component.position }
+	def positionToTopModifier: Vector2D = parent match
+	{
+		case Left(_) => Vector2D.zero
+		case Right((block, component)) => block.positionToTopModifier + component.position
+	}
 	
 	/**
 	  * @return A modifier used for calculating the absolute position of the bottom component (not on this hierarchy)
 	  */
-	def absolutePositionModifier = positionToTopModifier + top.canvas.absolutePosition
+	def absolutePositionModifier = positionToTopModifier + top.absolutePosition
 	
 	
 	// OTHER	--------------------------
 	
 	/**
-	  * Revalidates this component hierarchy
+	  * Revalidates this component hierarchy (provided this part of the hierarchy is currently linked to the main
+	  * stack hierarchy)
 	  */
-	def revalidate() =
+	@tailrec
+	final def revalidate(): Unit =
 	{
-		// Won't revalidate non-linked components
-		val targetComponents = blocks.reverseIterator.takeWhile { _.isLinked }.toVector
-		targetComponents.foreach { _.component.resetCachedSize() }
-		if (targetComponents.size == blocks.size)
-			top.canvas.revalidate()
+		if (isThisLevelLinked)
+			parent match
+			{
+				case Left(canvas) => canvas.revalidate()
+				case Right((block, component)) =>
+					component.resetCachedSize()
+					block.revalidate()
+			}
+	}
+	
+	/**
+	  * Revalidates this component hierarchy (provided this part of the hierarchy is currently linked to the main
+	  * stack hierarchy), then calls the specified function
+	  * @param f A function called once this hierarchy has been updated. Please note that this function might not
+	  *          get called at all.
+	  */
+	@tailrec
+	final def revalidateAndThen(f: => Unit): Unit =
+	{
+		if (isThisLevelLinked)
+			parent match
+			{
+				case Left(canvas) => canvas.revalidateAndThen(f)
+				case Right((block, component)) =>
+					component.resetCachedSize()
+					block.revalidateAndThen(f)
+			}
 	}
 	
 	/**
 	  * Repaints the whole component hierarchy (if linked)
 	  */
-	def repaintAll() = if (isLinked) top.canvas.repaint()
+	def repaintAll() = if (isLinked) top.repaint()
 	
 	/**
 	  * Repaints a sub-section of the bottom component (if linked to top)
@@ -73,16 +122,20 @@ class ComponentHierarchy(top: HierarchyTop, blocks: Vector[HierarchyBlock])
 	def repaint(area: => Bounds) =
 	{
 		if (isLinked)
-			top.canvas.repaint(area + positionToTopModifier)
+			top.repaint(area + positionToTopModifier)
 	}
 	
 	/**
 	  * Repaints the bottom component
 	  */
-	def repaintBottom() = blocks.lastOption match
+	def repaintBottom() =
 	{
-		case Some(block) => repaint(Bounds(Point.origin, block.component.size))
-		case None => repaintAll()
+		if (isLinked)
+			parent match
+			{
+				case Left(canvas) => canvas.repaint()
+				case Right((block, component)) => block.repaint(component.bounds)
+			}
 	}
 	
 	/**
@@ -90,56 +143,5 @@ class ComponentHierarchy(top: HierarchyTop, blocks: Vector[HierarchyBlock])
 	  * @return Metrics for that font
 	  */
 	// TODO: Refactor this once component class hierarchy has been updated
-	def fontMetrics(font: Font) = top.canvas.component.getFontMetrics(font.toAwt)
-	
-	/**
-	  * @param block A hierarchy block to be appended to the end of this hierarchy chain
-	  * @return A hierarchy with specified block included
-	  */
-	def +(block: HierarchyBlock) = new ComponentHierarchy(top, blocks :+ block)
-	
-	
-	// NESTED	--------------------------
-	
-	// Used for updating link status (whether this hierarchy is connected to the top)
-	private class LinkChangeListener
-	{
-		// ATTRIBUTES	------------------
-		
-		val pointer = new PointerWithEvents(false)
-		
-		private val blockListeners = blocks.flatMap { _.linkPointer }.map { p =>
-			val listener = new BlockListener(p.value)
-			p.addListener(listener)
-			listener
-		}
-		
-		
-		// INITIAL CODE	------------------
-		
-		top.attachedPointer.addListener { e => updateStatus(e.newValue) }
-		updateStatus()
-		
-		
-		// OTHER	----------------------
-		
-		private def updateStatus(): Unit = updateStatus(top.attachedPointer.value)
-		
-		private def updateStatus(isTopLinked: Boolean) =
-			pointer.value = isTopLinked && blockListeners.forall { _.linked }
-		
-		
-		// NESTED	----------------------
-		
-		class BlockListener(initialState: Boolean) extends ChangeListener[Boolean]
-		{
-			var linked = initialState
-			
-			override def onChangeEvent(event: ChangeEvent[Boolean]) =
-			{
-				linked = event.newValue
-				updateStatus()
-			}
-		}
-	}
+	def fontMetrics(font: Font) = top.component.getFontMetrics(font.toAwt)
 }